infer/torch/requirements.txt
pillow==8.3.2
timm==0.6.12